+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.2.13] - 2021-10-12
+### Added
+- Export new confirmation radio buttons component
+## [1.2.12] - 2021-10-12
+### Added
+- Added new confirmation radio buttons component
 ## [1.2.10] - 2021-10-06
 ### Added
 - Added new claim-line and checkout-with-text icons
@@ -145,7 +151,8 @@
 
 ### Changed
 - Updated gap and styles on Row component
-
+[1.2.13]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.12...v1.2.13
+[1.2.12]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.11...v1.2.12
 [1.2.10]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.9...v1.2.10
 [1.2.9]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.8...v1.2.9
 [1.2.8]: https://github.com/marshmallow-insurance/smores-react/compare/v1.2.7...v1.2.8